Use .gitignore as to document & exclude large assets
A section of .gitignore identifies the packages that must be imported to support the project, but also excludes them from the commit. 

With libraries, Unity is set up so that the library is added to the manifest.json to be automatically reimported, while the /Library folder itself can be ignored in commits. 

Unfortunately, other kinds of assets don't work this way, even though they are imported using the package manager and even though they are sometimes huge. Some of these are so bulky that I wasn't able to push them to GitHub.

This approach, I hope, will keep track of the assets that must be imported into the project manually and also ignore them in the commit, all in one place.