Unity Starter Assets - Third Person Character Controller
I'm including this one in the commit because scripts and inputactions are referenced directly from the scene. These things could be moved, but not copied without creating build conflicts. The namespaces could be change, and perhaps that is ultimately the answer. For now, I'll include it and reference it to avoid duplication.