Revert to Input System 1.3.0
The 1.4 versions of the Input System have a bug that causes error messages to spew into the console, even though everything's working fine. This makes it hard to see real error messages, so reverting to a version before the bug was introduced. 1.3.0 is the most recent version that is really usable.