Scaled down environment with hidden 'thing'
The player starts in SmallValley, in a terrain only 1000m across. Will this be big enough to accommodate the game? It is certainly easier to decorate!
Player can look for the purple box 'thing' near the starting point and pick it up.
Demonstrates that a small object will be hard to find in the environment without more hints.