Eternal Temple asset
Package from package manager installs in to
Assets/EternalTemple.
Finding the UDP package embedded in that one and double clicking installs the actual content into
Assets/Eternal Temple
Notice the space.
I'm putting it in .gitignore, so it will need to be manually installed if trying to make this project editable.