Nonphysics double of the thing
Rather than the wild, spinning following behavior of the "thing", I had intended it to look sort of normal, strapped to the back, for example. Doing this with the physics object has so far eluded me (even with joints, although at moments that did look good). 
This approach is less elegant, but it looks okay for now.
It is incomplete. It only really works for one pickup/drop. Easily fixed.